refactor: unify backend errors on AppError, drop dead frontend hooks
Backend: exercise/auth/user modules now throw typed AppErrors and
respond via sendError, matching the plan/workout convention. Replaces
error-message string matching in controllers. Status changes:
- PATCH/DELETE /api/exercises/:id not-owner: 400/500 -> 403
- GET /api/users/:id error shape: {message} -> {success,error},
  success shape now matches the documented {success,data} envelope
- GET /api/auth/me with valid token but deleted user stays 401
Docs updated (module API.md, openapi.yaml); E2E assertions unaffected.

Frontend: remove unused legacy fetch hooks (useWorkouts, useExercises,
useWorkout, ~660 lines) superseded by DataContext; Exercise type moved
to types/exercise.ts.
