+++ backend/src/modules/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 import * as userRepo from "../user/user.repository.js";
+import { ConflictError, UnauthorizedError } from "../../common/errors.js";
 import type { RegisterDto, LoginDto } from "./auth.schema.js";
 
 const JWT_SECRET = process.env.JWT_SECRET;
@@ -9,7 +10,7 @@ const JWT_EXPIRES_IN = "7d";
 export const register = async (data: RegisterDto) => {
   const existingUser = await userRepo.findUserByEmail(data.email);
   if (existingUser) {
-    throw new Error("Użytkownik z tym emailem już istnieje");
+    throw new ConflictError("Użytkownik z tym emailem już istnieje");
   }
 
   const hashedPassword = await bcrypt.hash(data.password, 10);
@@ -40,12 +41,12 @@ export const register = async (data: RegisterDto) => {
 export const login = async (data: LoginDto) => {
   const user = await userRepo.findUserByEmail(data.email);
   if (!user) {
-    throw new Error("Nieprawidłowy email lub hasło");
+    throw new UnauthorizedError("Nieprawidłowy email lub hasło");
   }
 
   const isPasswordValid = await bcrypt.compare(data.password, user.password);
   if (!isPasswordValid) {
-    throw new Error("Nieprawidłowy email lub hasło");
+    throw new UnauthorizedError("Nieprawidłowy email lub hasło");
   }
 
   if (!JWT_SECRET) {
@@ -73,7 +74,7 @@ export const verifyToken = (token: string) => {
   try {
     return jwt.verify(token, JWT_SECRET) as { userId: string; email: string };
   } catch (error) {
-    throw new Error("Nieprawidłowy lub wygasły token");
+    throw new UnauthorizedError("Nieprawidłowy lub wygasły token");
   }
 };
 
@@ -81,7 +82,8 @@ export const getUserFromToken = async (token: string) => {
   const decoded = verifyToken(token);
   const user = await userRepo.findUserById(decoded.userId);
   if (!user) {
-    throw new Error("Użytkownik nie znaleziony");
+    // 401 (nie 404): ważny token bez istniejącego użytkownika ma wymusić wylogowanie
+    throw new UnauthorizedError("Użytkownik nie znaleziony");
   }
   const { password: _, ...userWithoutPassword } = user;
   return userWithoutPassword;
