+++ backend/src/modules/user/user.service.ts
@@ -1,9 +1,10 @@
+import { NotFoundError } from "../../common/errors.js";
 import { findUserById } from "./user.repository.js";
 
 export const getUserById = async (id: string) => {
   const user = await findUserById(id);
   if (!user) {
-    throw new Error("User not found");
+    throw new NotFoundError("User not found");
   }
   const { password, ...userWithoutPassword } = user;
   return userWithoutPassword;
