+++ frontend/src/components/screens/ExercisesScreen.tsx
@@ -1,7 +1,7 @@
 import { memo } from "react";
 import { useData } from "@/contexts/data";
 import { ExerciseList } from "../exercises/ExerciseList";
-import type { Exercise } from "@/hooks/useExercises";
+import type { Exercise } from "@/types";
 
 interface ExercisesScreenProps {
   onAddExercise: () => void;
