+++ frontend/src/hooks/index.ts
@@ -1,11 +1,2 @@
 export { useNavigation } from "./useNavigation";
 export { useTheme } from "./useTheme";
-export { useExercises } from "./useExercises";
-export {
-  useWorkouts,
-  useWorkout,
-  useExerciseStats,
-  useAllUserStats,
-  useActiveWorkout,
-} from "./useWorkouts";
-export type { ExerciseStats } from "@/types";
