--- frontend/src/hooks/useWorkout.ts
@@ -1,13 +1,0 @@
-import { useCallback } from "react";
-
-interface UseWorkoutReturn {
-  startAddWorkout: () => void;
-}
-
-export function useWorkout(): UseWorkoutReturn {
-  const startAddWorkout = useCallback(() => {}, []);
-
-  return {
-    startAddWorkout,
-  };
-}
