+++ frontend/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './navigation'
 export * from './user'
 export * from './workout'
+export * from './exercise'
